COLONY: use int64 for some variables that can be large when calculating projects

+++ engines/colony/render.cpp
@@ -317,7 +317,7 @@ bool isSurfaceVisible(const int *surface, int pointCount, const int *screenX, co
 				if (dyp < 0)
 					return true;
 			} else {
-				const int32 b = dy * dxp - dx * dyp;
+				const int64 b = (int64)dy * dxp - (int64)dx * dyp;
 				if (b > 0)
 					return false;
 				if (b < 0)
@@ -330,7 +330,7 @@ bool isSurfaceVisible(const int *surface, int pointCount, const int *screenX, co
 				if (dyp > 0)
 					return true;
 			} else {
-				const int32 b = dx * dyp - dy * dxp;
+				const int64 b = (int64)dx * dyp - (int64)dy * dxp;
 				if (b < 0)
 					return false;
 				if (b > 0)
